Move routes to Router component
