+++ pet-clinic/src/components/Router.jsx
@@ -1,6 +1,8 @@
 import { useRoutes } from "react-router-dom";
 import LoginUser from "../views/LoginUser";
-import App from "../App";
+import { useLocation, useNavigate } from "react-router-dom";
+import { useEffect } from "react";
+import { useSelector } from "react-redux";
 import CONST from "../constants/index";
 import RegisterUser from "../views/RegisterUser";
 import Dashboard from "../Dashboard";
@@ -14,66 +16,81 @@ import Disease from "../views/Disease/Disease";
 import AddDisease from "../views/Disease/AddDisease";
 import EditDisease from "../views/Disease/EditDisease";
 import User from "../views/User/User";
+import AddUser from "../views/User/AddUser";
 
 function Router() {
+  const location = useLocation();
+  const navigate = useNavigate();
+
+  const token = useSelector((state) => {
+    return state.authReducer.token;
+  });
+
+  useEffect(() => {
+    if (
+      location.pathname.split("/").includes(CONST.ROUTE.DASHBOARD) &&
+      !token
+    ) {
+      navigate(`/${CONST.ROUTE.LOGIN}`);
+    }
+  }, [location]);
+
   return useRoutes([
     {
-      path: "/",
-      element: <App />,
+      path: CONST.ROUTE.LOGIN,
+      element: <LoginUser />,
+    },
+    {
+      path: CONST.ROUTE.REGISTER,
+      element: <RegisterUser />,
+    },
+    {
+      path: `/${CONST.ROUTE.DASHBOARD}`,
+      element: <Dashboard />,
       children: [
         {
-          path: CONST.ROUTE.LOGIN,
-          element: <LoginUser />,
+          path: CONST.ROUTE.ANIMAL_TYPE,
+          element: <AnimalType />,
         },
         {
-          path: CONST.ROUTE.REGISTER,
-          element: <RegisterUser />,
+          path: `${CONST.ROUTE.ANIMAL_TYPE}/${CONST.ROUTE.ADD}`,
+          element: <AddAnimalType />,
         },
         {
-          path: `/${CONST.ROUTE.DASHBOARD}`,
-          element: <Dashboard />,
-          children: [
-            {
-              path: CONST.ROUTE.ANIMAL_TYPE,
-              element: <AnimalType />,
-            },
-            {
-              path: `${CONST.ROUTE.ANIMAL_TYPE}/${CONST.ROUTE.ADD}`,
-              element: <AddAnimalType />,
-            },
-            {
-              path: `${CONST.ROUTE.ANIMAL_TYPE}/:id/${CONST.ROUTE.EDIT}`,
-              element: <EditAnimalType />,
-            },
-            {
-              path: CONST.ROUTE.ANIMAL,
-              element: <Animal />,
-            },
-            {
-              path: `${CONST.ROUTE.ANIMAL}/${CONST.ROUTE.ADD}`,
-              element: <AddAnimal />,
-            },
-            {
-              path: `${CONST.ROUTE.ANIMAL}/:id/${CONST.ROUTE.EDIT}`,
-              element: <EditAnimal />,
-            },
-            {
-              path: CONST.ROUTE.DISEASE,
-              element: <Disease />,
-            },
-            {
-              path: `${CONST.ROUTE.DISEASE}/${CONST.ROUTE.ADD}`,
-              element: <AddDisease />,
-            },
-            {
-              path: `${CONST.ROUTE.DISEASE}/:id/${CONST.ROUTE.EDIT}`,
-              element: <EditDisease />,
-            },
-            {
-              path: CONST.ROUTE.USER,
-              element: <User />,
-            },
-          ],
+          path: `${CONST.ROUTE.ANIMAL_TYPE}/:id/${CONST.ROUTE.EDIT}`,
+          element: <EditAnimalType />,
+        },
+        {
+          path: CONST.ROUTE.ANIMAL,
+          element: <Animal />,
+        },
+        {
+          path: `${CONST.ROUTE.ANIMAL}/${CONST.ROUTE.ADD}`,
+          element: <AddAnimal />,
+        },
+        {
+          path: `${CONST.ROUTE.ANIMAL}/:id/${CONST.ROUTE.EDIT}`,
+          element: <EditAnimal />,
+        },
+        {
+          path: CONST.ROUTE.DISEASE,
+          element: <Disease />,
+        },
+        {
+          path: `${CONST.ROUTE.DISEASE}/${CONST.ROUTE.ADD}`,
+          element: <AddDisease />,
+        },
+        {
+          path: `${CONST.ROUTE.DISEASE}/:id/${CONST.ROUTE.EDIT}`,
+          element: <EditDisease />,
+        },
+        {
+          path: CONST.ROUTE.USER,
+          element: <User />,
+        },
+        {
+          path: `${CONST.ROUTE.USER}/${CONST.ROUTE.ADD}`,
+          element: <AddUser />,
         },
       ],
     },
